HistoryToken static methods pulled from SpreadsheetHistoryToken
- No code/logic changes.
