+++ src/main/java/walkingkooka/spreadsheet/dominokit/history/HistoryToken.java
@@ -20,7 +20,16 @@ package walkingkooka.spreadsheet.dominokit.history;
 import walkingkooka.net.HasUrlFragment;
 import walkingkooka.net.UrlFragment;
 import walkingkooka.predicate.character.CharPredicates;
+import walkingkooka.spreadsheet.SpreadsheetFormula;
+import walkingkooka.spreadsheet.SpreadsheetId;
+import walkingkooka.spreadsheet.SpreadsheetName;
 import walkingkooka.spreadsheet.dominokit.AppContext;
+import walkingkooka.spreadsheet.format.pattern.SpreadsheetPattern;
+import walkingkooka.spreadsheet.format.pattern.SpreadsheetPatternKind;
+import walkingkooka.spreadsheet.meta.SpreadsheetMetadataPropertyName;
+import walkingkooka.spreadsheet.reference.SpreadsheetLabelMapping;
+import walkingkooka.spreadsheet.reference.SpreadsheetLabelName;
+import walkingkooka.spreadsheet.reference.SpreadsheetViewportSelection;
 import walkingkooka.text.CharSequences;
 import walkingkooka.text.cursor.TextCursor;
 import walkingkooka.text.cursor.TextCursorSavePoint;
@@ -30,6 +39,7 @@ import walkingkooka.text.cursor.parser.ParserContext;
 import walkingkooka.text.cursor.parser.ParserContexts;
 import walkingkooka.text.cursor.parser.Parsers;
 import walkingkooka.text.cursor.parser.StringParserToken;
+import walkingkooka.tree.text.TextStylePropertyName;
 
 import java.util.Objects;
 import java.util.Optional;
@@ -37,6 +47,465 @@ import java.util.Optional;
 public abstract class HistoryToken implements HasUrlFragment {
 
     private final static int MAX_LENGTH = 8192;
+
+    /**
+     * {@see SpreadsheetCellSelectHistoryToken}
+     */
+    public static SpreadsheetCellSelectHistoryToken cell(final SpreadsheetId id,
+                                                         final SpreadsheetName name,
+                                                         final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetCellSelectHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetCellFormulaSelectHistoryToken}
+     */
+    public static SpreadsheetCellFormulaSelectHistoryToken formula(final SpreadsheetId id,
+                                                                   final SpreadsheetName name,
+                                                                   final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetCellFormulaSelectHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetCellFormulaSaveHistoryToken}
+     */
+    public static SpreadsheetCellFormulaSaveHistoryToken formulaSave(final SpreadsheetId id,
+                                                                     final SpreadsheetName name,
+                                                                     final SpreadsheetViewportSelection viewportSelection,
+                                                                     final SpreadsheetFormula formula) {
+        return SpreadsheetCellFormulaSaveHistoryToken.with(
+                id,
+                name,
+                viewportSelection,
+                formula
+        );
+    }
+
+    /**
+     * {@see SpreadsheetCellClearHistoryToken}
+     */
+    public static SpreadsheetCellClearHistoryToken cellClear(final SpreadsheetId id,
+                                                             final SpreadsheetName name,
+                                                             final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetCellClearHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetCellDeleteHistoryToken}
+     */
+    public static SpreadsheetCellDeleteHistoryToken cellDelete(final SpreadsheetId id,
+                                                               final SpreadsheetName name,
+                                                               final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetCellDeleteHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetCellFreezeHistoryToken}
+     */
+    public static SpreadsheetCellFreezeHistoryToken cellFreeze(final SpreadsheetId id,
+                                                               final SpreadsheetName name,
+                                                               final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetCellFreezeHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetCellMenuHistoryToken}
+     */
+    public static SpreadsheetCellMenuHistoryToken cellMenu(final SpreadsheetId id,
+                                                           final SpreadsheetName name,
+                                                           final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetCellMenuHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetCellPatternSelectHistoryToken}
+     */
+    public static SpreadsheetCellPatternSelectHistoryToken cellPattern(final SpreadsheetId id,
+                                                                       final SpreadsheetName name,
+                                                                       final SpreadsheetViewportSelection viewportSelection,
+                                                                       final SpreadsheetPatternKind patternKind) {
+        return SpreadsheetCellPatternSelectHistoryToken.with(
+                id,
+                name,
+                viewportSelection,
+                patternKind
+        );
+    }
+
+    /**
+     * {@see SpreadsheetCellPatternSaveHistoryToken}
+     */
+    public static SpreadsheetCellPatternSaveHistoryToken cellPatternSave(final SpreadsheetId id,
+                                                                         final SpreadsheetName name,
+                                                                         final SpreadsheetViewportSelection viewportSelection,
+                                                                         final SpreadsheetPatternKind patternKind,
+                                                                         final Optional<SpreadsheetPattern> pattern) {
+        return SpreadsheetCellPatternSaveHistoryToken.with(
+                id,
+                name,
+                viewportSelection,
+                patternKind,
+                pattern
+        );
+    }
+
+    /**
+     * {@see SpreadsheetCellStyleSelectHistoryToken}
+     */
+    public static <T> SpreadsheetCellStyleSelectHistoryToken<T> cellStyle(final SpreadsheetId id,
+                                                                          final SpreadsheetName name,
+                                                                          final SpreadsheetViewportSelection viewportSelection,
+                                                                          final TextStylePropertyName<T> propertyName) {
+        return SpreadsheetCellStyleSelectHistoryToken.with(
+                id,
+                name,
+                viewportSelection,
+                propertyName
+        );
+    }
+
+    /**
+     * {@see SpreadsheetCellStyleSaveHistoryToken}
+     */
+    public static <T> SpreadsheetCellStyleSaveHistoryToken<T> cellStyleSave(final SpreadsheetId id,
+                                                                            final SpreadsheetName name,
+                                                                            final SpreadsheetViewportSelection viewportSelection,
+                                                                            final TextStylePropertyName<T> propertyName,
+                                                                            final Optional<T> propertyValue) {
+        return SpreadsheetCellStyleSaveHistoryToken.with(
+                id,
+                name,
+                viewportSelection,
+                propertyName,
+                propertyValue
+        );
+    }
+
+    /**
+     * {@see SpreadsheetCellUnfreezeHistoryToken}
+     */
+    public static SpreadsheetCellUnfreezeHistoryToken cellUnfreeze(final SpreadsheetId id,
+                                                                   final SpreadsheetName name,
+                                                                   final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetCellUnfreezeHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetColumnSelectHistoryToken}
+     */
+    public static SpreadsheetColumnSelectHistoryToken column(final SpreadsheetId id,
+                                                             final SpreadsheetName name,
+                                                             final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetColumnSelectHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetColumnClearHistoryToken}
+     */
+    public static SpreadsheetColumnClearHistoryToken columnClear(final SpreadsheetId id,
+                                                                 final SpreadsheetName name,
+                                                                 final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetColumnClearHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetColumnDeleteHistoryToken}
+     */
+    public static SpreadsheetColumnDeleteHistoryToken columnDelete(final SpreadsheetId id,
+                                                                   final SpreadsheetName name,
+                                                                   final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetColumnDeleteHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetColumnFreezeHistoryToken}
+     */
+    public static SpreadsheetColumnFreezeHistoryToken columnFreeze(final SpreadsheetId id,
+                                                                   final SpreadsheetName name,
+                                                                   final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetColumnFreezeHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetColumnMenuHistoryToken}
+     */
+    public static SpreadsheetColumnMenuHistoryToken columnMenu(final SpreadsheetId id,
+                                                               final SpreadsheetName name,
+                                                               final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetColumnMenuHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetColumnUnfreezeHistoryToken}
+     */
+    public static SpreadsheetColumnUnfreezeHistoryToken columnUnfreeze(final SpreadsheetId id,
+                                                                       final SpreadsheetName name,
+                                                                       final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetColumnUnfreezeHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetLabelMappingSelectHistoryToken}
+     */
+    public static SpreadsheetLabelMappingSelectHistoryToken labelMapping(final SpreadsheetId id,
+                                                                         final SpreadsheetName name,
+                                                                         final SpreadsheetLabelName label) {
+        return SpreadsheetLabelMappingSelectHistoryToken.with(
+                id,
+                name,
+                label
+        );
+    }
+
+    /**
+     * {@see SpreadsheetLabelMappingDeleteHistoryToken}
+     */
+    public static SpreadsheetLabelMappingDeleteHistoryToken labelMappingDelete(final SpreadsheetId id,
+                                                                               final SpreadsheetName name,
+                                                                               final SpreadsheetLabelName labelName) {
+        return SpreadsheetLabelMappingDeleteHistoryToken.with(
+                id,
+                name,
+                labelName
+        );
+    }
+
+    /**
+     * {@see SpreadsheetLabelMappingSaveHistoryToken}
+     */
+    public static SpreadsheetLabelMappingSaveHistoryToken labelMappingSave(final SpreadsheetId id,
+                                                                           final SpreadsheetName name,
+                                                                           final SpreadsheetLabelMapping mapping) {
+        return SpreadsheetLabelMappingSaveHistoryToken.with(
+                id,
+                name,
+                mapping
+        );
+    }
+
+    /**
+     * {@see SpreadsheetMetadataPropertySelectHistoryToken}
+     */
+    public static <T> SpreadsheetMetadataPropertySelectHistoryToken<T> metadataPropertySelect(final SpreadsheetId id,
+                                                                                              final SpreadsheetName name,
+                                                                                              final SpreadsheetMetadataPropertyName<T> propertyName) {
+        return SpreadsheetMetadataPropertySelectHistoryToken.with(
+                id,
+                name,
+                propertyName
+        );
+    }
+
+    /**
+     * {@see SpreadsheetMetadataPropertySaveHistoryToken}
+     */
+    public static <T> SpreadsheetMetadataPropertySaveHistoryToken<T> metadataPropertySave(final SpreadsheetId id,
+                                                                                          final SpreadsheetName name,
+                                                                                          final SpreadsheetMetadataPropertyName<T> propertyName,
+                                                                                          final Optional<T> propertyValue) {
+        return SpreadsheetMetadataPropertySaveHistoryToken.with(
+                id,
+                name,
+                propertyName,
+                propertyValue
+        );
+    }
+
+    /**
+     * {@see SpreadsheetMetadataPropertyStyleSelectHistoryToken}
+     */
+    public static <T> SpreadsheetMetadataPropertyStyleSelectHistoryToken<T> metadataPropertyStyle(final SpreadsheetId id,
+                                                                                                  final SpreadsheetName name,
+                                                                                                  final TextStylePropertyName<T> stylePropertyName) {
+        return SpreadsheetMetadataPropertyStyleSelectHistoryToken.with(
+                id,
+                name,
+                stylePropertyName
+        );
+    }
+
+    /**
+     * {@see SpreadsheetMetadataPropertyStyleSaveHistoryToken}
+     */
+    public static <T> SpreadsheetMetadataPropertyStyleSaveHistoryToken<T> metadataPropertyStyleSave(final SpreadsheetId id,
+                                                                                                    final SpreadsheetName name,
+                                                                                                    final TextStylePropertyName<T> stylePropertyName,
+                                                                                                    final Optional<T> stylePropertyValue) {
+        return SpreadsheetMetadataPropertyStyleSaveHistoryToken.with(
+                id,
+                name,
+                stylePropertyName,
+                stylePropertyValue
+        );
+    }
+
+    /**
+     * {@see SpreadsheetMetadataSelectHistoryToken}
+     */
+    public static SpreadsheetMetadataSelectHistoryToken metadataSelect(final SpreadsheetId id,
+                                                                       final SpreadsheetName name) {
+        return SpreadsheetMetadataSelectHistoryToken.with(
+                id,
+                name
+        );
+    }
+
+    /**
+     * {@see SpreadsheetRowSelectHistoryToken}
+     */
+    public static SpreadsheetRowSelectHistoryToken row(final SpreadsheetId id,
+                                                       final SpreadsheetName name,
+                                                       final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetRowSelectHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetRowClearHistoryToken}
+     */
+    public static SpreadsheetRowClearHistoryToken rowClear(final SpreadsheetId id,
+                                                           final SpreadsheetName name,
+                                                           final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetRowClearHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetRowDeleteHistoryToken}
+     */
+    public static SpreadsheetRowDeleteHistoryToken rowDelete(final SpreadsheetId id,
+                                                             final SpreadsheetName name,
+                                                             final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetRowDeleteHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetRowFreezeHistoryToken}
+     */
+    public static SpreadsheetRowFreezeHistoryToken rowFreeze(final SpreadsheetId id,
+                                                             final SpreadsheetName name,
+                                                             final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetRowFreezeHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetRowMenuHistoryToken}
+     */
+    public static SpreadsheetRowMenuHistoryToken rowMenu(final SpreadsheetId id,
+                                                         final SpreadsheetName name,
+                                                         final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetRowMenuHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetRowUnfreezeHistoryToken}
+     */
+    public static SpreadsheetRowUnfreezeHistoryToken rowUnfreeze(final SpreadsheetId id,
+                                                                 final SpreadsheetName name,
+                                                                 final SpreadsheetViewportSelection viewportSelection) {
+        return SpreadsheetRowUnfreezeHistoryToken.with(
+                id,
+                name,
+                viewportSelection
+        );
+    }
+
+    /**
+     * {@see SpreadsheetCreateHistoryToken}
+     */
+    public static SpreadsheetCreateHistoryToken spreadsheetCreate() {
+        return SpreadsheetCreateHistoryToken.with();
+    }
+
+    /**
+     * {@see SpreadsheetLoadHistoryToken}
+     */
+    public static SpreadsheetLoadHistoryToken spreadsheetLoad(final SpreadsheetId id) {
+        return SpreadsheetLoadHistoryToken.with(
+                id
+        );
+    }
+
+    /**
+     * {@see SpreadsheetSelectHistoryToken}
+     */
+    public static SpreadsheetSelectHistoryToken spreadsheetSelect(final SpreadsheetId id,
+                                                                  final SpreadsheetName name) {
+        return SpreadsheetSelectHistoryToken.with(
+                id,
+                name
+        );
+    }
 
     /**
      * {@see UnknownHistoryToken}
@@ -60,14 +529,14 @@ public abstract class HistoryToken implements HasUrlFragment {
                 final char c = cursor.at();
                 switch (c) {
                     case '/':
-                        token = SpreadsheetHistoryToken.spreadsheetCreate()
+                        token = HistoryToken.spreadsheetCreate()
                                 .parse(cursor);
                         break;
                     default:
                         token = UnknownHistoryToken.with(fragment);
                 }
             } else {
-                token = SpreadsheetHistoryToken.spreadsheetCreate();
+                token = HistoryToken.spreadsheetCreate();
             }
         } catch (final RuntimeException ignore) {
             token = UnknownHistoryToken.with(fragment);
